+++ src/sensors-sys.h
@@ -1,2 +1,0 @@
-#include <sensors/sensors.h>
-#include <sensors/error.h>