+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.1] - 2021-10-25
+
+### Changed
+
+- Added a *Documentation-only* build mode to allow building documentation even
+  if `libsensors` and its headers are unavailable.
+  This allows building on `docs.rs`.
+
 ## [0.2.0] - 2021-10-25
 
 ### Changed